refactor(data-exports): move source size calculation into DataExport
Fold the one-caller size calculator into the model and authorize selected
IDs before running the source-size limit query.
